Add some new features. Configuration files of openCV are included.
Features(refer to Bells and Whistles in project page):
1.One problem with the live wire is that it prefers shorter paths so will tend to cut through large object rather than wrap around them.
One way to fix this is specify a specific region in which the path must stay.
As long as this region contains the object boundary but excludes most of the interior, the path will be forced to follow the boundary.
One way of specifying such a region is to use a thick (e.g., 50 pixel wide) paint brush. Implement this feature.
This feature is provided by setMask. You need specify the region by mask before setSeed

2.Try different costs functions, for example the method described in Intelligent Scissors for Image Composition, and modify the user interface to allow the user to select different functions.
This feature is provided by setCostFunction. In the construction function of Iscissor, you can specify the cost function.

3.The only point that doesn't snap to edges is the seed. Implement a seed snapping feature, where the seed is automatically moved to the closest edge.
This feature is provided by snapSeed. The function will return the new seed and meanWhile setSeed and update path tree.